Refactor workhorse functions from `cli.py` to `engine.py`
