+++ calypso_anemometer/cli.py
@@ -2,18 +2,15 @@
 # (c) 2022 Andreas Motl <andreas.motl@panodata.org>
 # License: GNU Affero General Public License, Version 3
 import logging
-import os
-import sys
-from typing import Callable, Optional
+import typing as t
 
 import click
 
 from calypso_anemometer.core import CalypsoDeviceApi
-from calypso_anemometer.exception import CalypsoError
+from calypso_anemometer.engine import handler_factory, run_engine
 from calypso_anemometer.fake import CalypsoDeviceApiFake
-from calypso_anemometer.model import CalypsoDeviceDataRate, CalypsoDeviceMode, CalypsoReading
-from calypso_anemometer.telemetry import TelemetryAdapter
-from calypso_anemometer.util import EnumChoice, make_sync, setup_logging, wait_forever
+from calypso_anemometer.model import CalypsoDeviceDataRate, CalypsoDeviceMode
+from calypso_anemometer.util import EnumChoice, make_sync, setup_logging
 
 logger = logging.getLogger(__name__)
 
@@ -31,14 +28,14 @@ def cli(ctx, verbose, debug):
 @click.pass_context
 @make_sync
 async def info(ctx):
-    await calypso_run(lambda calypso: calypso.about())
+    await run_engine(workhorse=CalypsoDeviceApi, handler=lambda calypso: calypso.about())
 
 
 @click.command()
 @click.pass_context
 @make_sync
 async def explore(ctx):
-    await calypso_run(lambda calypso: calypso.explore())
+    await run_engine(workhorse=CalypsoDeviceApi, handler=lambda calypso: calypso.explore())
 
 
 rate_option = click.option(
@@ -61,7 +58,7 @@ target_option = click.option("--target", type=str, required=False, help="Submit 
 @rate_option
 @click.pass_context
 @make_sync
-async def set_option(ctx, mode: Optional[CalypsoDeviceMode] = None, rate: Optional[CalypsoDeviceDataRate] = None):
+async def set_option(ctx, mode: t.Optional[CalypsoDeviceMode] = None, rate: t.Optional[CalypsoDeviceDataRate] = None):
     async def handler(calypso: CalypsoDeviceApi):
         if mode is not None:
             logger.info(f"Setting device mode to {mode}")
@@ -71,7 +68,7 @@ async def set_option(ctx, mode: Optional[CalypsoDeviceMode] = None, rate: Option
             await calypso.set_datarate(rate)
         await calypso.about()
 
-    await calypso_run(handler)
+    await run_engine(workhorse=CalypsoDeviceApi, handler=handler)
 
 
 @click.command()
@@ -81,10 +78,10 @@ async def set_option(ctx, mode: Optional[CalypsoDeviceMode] = None, rate: Option
 @click.pass_context
 @make_sync
 async def read(
-    ctx, subscribe: bool = False, target: Optional[str] = None, rate: Optional[CalypsoDeviceDataRate] = None
+    ctx, subscribe: bool = False, target: t.Optional[str] = None, rate: t.Optional[CalypsoDeviceDataRate] = None
 ):
     handler = await handler_factory(subscribe=subscribe, target=target, rate=rate)
-    await calypso_run(handler)
+    await run_engine(workhorse=CalypsoDeviceApi, handler=handler)
 
 
 @click.command()
@@ -94,59 +91,10 @@ async def read(
 @click.pass_context
 @make_sync
 async def fake(
-    ctx, subscribe: bool = False, target: Optional[str] = None, rate: Optional[CalypsoDeviceDataRate] = None
+    ctx, subscribe: bool = False, target: t.Optional[str] = None, rate: t.Optional[CalypsoDeviceDataRate] = None
 ):
     handler = await handler_factory(subscribe=subscribe, target=target, rate=rate)
-    await fake_run(handler)
-
-
-async def calypso_run(callback: Callable):
-    try:
-        async with CalypsoDeviceApi(ble_address=os.getenv("CALYPSO_ADDRESS")) as calypso:
-            await callback(calypso)
-    except CalypsoError as ex:
-        logger.error(ex)
-        sys.exit(1)
-
-
-async def fake_run(callback: Callable):
-    try:
-        async with CalypsoDeviceApiFake(ble_address=os.getenv("CALYPSO_ADDRESS")) as calypso:
-            await callback(calypso)
-    except CalypsoError as ex:
-        logger.error(ex)
-        sys.exit(1)
-
-
-async def handler_factory(
-    subscribe: bool = False, target: Optional[str] = None, rate: Optional[CalypsoDeviceDataRate] = None
-):
-    telemetry = None
-    if target is not None:
-        telemetry = TelemetryAdapter(uri=target)
-
-    async def handler(calypso: CalypsoDeviceApi):
-
-        # One-shot reading.
-        if not subscribe:
-            reading = await calypso.get_reading()
-            reading.print()
-
-        # Continuous readings.
-        else:
-            if rate is not None:
-                logger.info(f"Setting device data rate to {rate}")
-                await calypso.set_datarate(rate)
-
-            def process_reading(reading: CalypsoReading):
-                reading.print()
-                if telemetry is not None:
-                    telemetry.submit(reading)
-
-            await calypso.subscribe_reading(process_reading)
-            await wait_forever()
-
-    return handler
+    await run_engine(workhorse=CalypsoDeviceApiFake, handler=handler)
 
 
 cli.add_command(info, name="info")
